[NCL-4958] moved internal endpoints to the module rest-new and deleted unused endpoint that made problems
