--- rest-model/src/main/java/org/jboss/pnc/rest/restmodel/bpm/BpmTaskRest.java
@@ -1,55 +1,0 @@
-/**
- * JBoss, Home of Professional Open Source.
- * Copyright 2014-2019 Red Hat, Inc., and individual contributors
- * as indicated by the @author tags.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.jboss.pnc.rest.restmodel.bpm;
-
-import lombok.*;
-
-import java.util.List;
-
-/**
- * @author Jakub Senko
- */
-@ToString
-@AllArgsConstructor
-public class BpmTaskRest {
-
-    /**
-     * Id of the task within BpmManager.
-     */
-    @Getter
-    @Setter
-    private Integer taskId;
-
-    /**
-     * Id of the process instance within BPM engine.
-     */
-    @Getter
-    @Setter
-    private Long processInstanceId;
-
-    /**
-     * Name of the process executing this task.
-     */
-    @Getter
-    @Setter
-    private String processName;
-
-    @Getter
-    @Setter
-    private List<BpmNotificationRest> events;
-}
