+++ rest-new/src/main/java/org/jboss/pnc/rest/endpoints/internal/BpmEndpoint.java
@@ -1,0 +1,56 @@
+/**
+ * JBoss, Home of Professional Open Source.
+ * Copyright 2014-2019 Red Hat, Inc., and individual contributors
+ * as indicated by the @author tags.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.jboss.pnc.rest.endpoints.internal;
+
+
+import javax.ws.rs.Consumes;
+import javax.ws.rs.Path;
+import javax.ws.rs.Produces;
+import javax.ws.rs.core.MediaType;
+
+import static org.jboss.pnc.rest.configuration.SwaggerConstants.SUCCESS_CODE;
+
+import javax.ws.rs.POST;
+import javax.ws.rs.PathParam;
+
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.tags.Tag;
+
+/**
+ * This endpoint is used for starting and interacting with BPM processes.
+ *
+ * @author Jakub Senko
+ */
+@Tag(name = "Internal")
+@Path("/bpm")
+@Produces(MediaType.APPLICATION_JSON)
+@Consumes(MediaType.APPLICATION_JSON)
+public interface BpmEndpoint{
+
+    @Operation(summary = "Notify PNC about a BPM task event. Accepts polymorphic JSON {\"eventType\": \"string\"} based on \"eventType\" field.",
+            responses = {
+                @ApiResponse(responseCode = SUCCESS_CODE, description = "Success")
+    })
+    @POST
+    @Path("/tasks/{taskId}/notify")
+    @Consumes(MediaType.APPLICATION_JSON)
+    void notifyTask(@Parameter(description = "BPM task ID") @PathParam("taskId") int taskId);
+
+}
